+++ src/main/java/com/sms/domain/User.java
@@ -11,7 +11,7 @@ public class User {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column
-	private Long userId;
+	private Integer userId;
 	
 	@Column
 	private String userEmail;
@@ -20,13 +20,13 @@ public class User {
 	private String userPassword;
 	
 	@Column
-	private Long roleId;
+	private Integer roleId;
 
-	public Long getUserId() {
+	public Integer getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Long userId) {
+	public void setUserId(Integer userId) {
 		this.userId = userId;
 	}
 
@@ -42,11 +42,11 @@ public class User {
 		this.userPassword = userPassword;
 	}
 
-	public Long getRoleId() {
+	public Integer getRoleId() {
 		return roleId;
 	}
 
-	public void setRoleId(Long roleId) {
+	public void setRoleId(Integer roleId) {
 		this.roleId = roleId;
 	}
 }
