+++ src/main/java/com/sms/repo/LocalityRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sms.domain.Locality;
 
-public interface LocalityRepository extends CrudRepository<Locality, Long> {
+public interface LocalityRepository extends CrudRepository<Locality, Integer> {
 	
-	public List<Locality> findLocalityByPincodeIdOrderByLocalityNameAsc(Long id);
+	public List<Locality> findLocalityByPincodeIdOrderByLocalityNameAsc(Integer id);
 
 }
