+++ src/main/java/com/sms/repo/PincodeRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sms.domain.Pincode;
 
-public interface PincodeRepository extends CrudRepository<Pincode, Long>{
+public interface PincodeRepository extends CrudRepository<Pincode, Integer>{
 	
-	public List<Pincode> findPincodeByCityIdOrderByPincodeNameAsc(Long id);
+	public List<Pincode> findPincodeByCityIdOrderByPincodeNameAsc(Integer id);
 	
 }
