+++ src/main/java/com/sms/repo/SocietyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sms.domain.Society;
 
-public interface SocietyRepository extends CrudRepository<Society, Long> {
-	public Society findBySocietyNameAndLocalityId(String societyName, Long localityName);
+public interface SocietyRepository extends CrudRepository<Society, Integer> {
+	public Society findBySocietyNameAndLocalityId(String societyName, Integer localityName);
 }
