+++ src/main/java/com/sms/service/SocietyService.java
@@ -3,7 +3,7 @@ package com.sms.service;
 import com.sms.domain.Society;
 
 public interface SocietyService {
-	public Society search(String societyName, Long localityId);
+	public Society search(String societyName, Integer localityId);
 	
 	public boolean register(Society society);
 }
