+++ src/main/java/com/sms/service/StateService.java
@@ -8,8 +8,8 @@ import com.sms.domain.State;
 public interface StateService {
 	
 	public List<Map<String, String>> getAll();
-	public List<State> getByCountryId(Long countryId);
+	public List<State> getByCountryId(Integer countryId);
 	public State save(State state);
-	public State getByStateId(Long stateId);
+	public State getByStateId(Integer stateId);
 	
 }
